dont send 400 for bad reports
only try to package valid reports. This is a safety measure and not for feedback.

Creating reports is happening asynchronously, so it has no use to return a 400 for bad reports
as it is, the caller can not act on it anyway. The bad report should have been flagged upon creation/update.